refactor(user-managemente): move search from filters to queries
